# docs/openapi/knowledge.openapi.yaml
openapi: 3.1.0
info:
  title: knowledge
  version: 1.0.0
  description: knowledge 控制面、知识生产、发布、检索与 Agent Run API。
paths:
...
  /kbs/{kb_id}/tasks/import:
    post:
      tags:
      - ingestion_tasks
      summary: Create Import Task
      operationId: create_import_task_kbs__kb_id__tasks_import_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/tasks/reindex:
    post:
      tags:
      - ingestion_tasks
      summary: Create Reindex Task
      operationId: create_reindex_task_kbs__kb_id__tasks_reindex_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/tasks/delete:
    post:
      tags:
      - ingestion_tasks
      summary: Create Delete Task
      operationId: create_delete_task_kbs__kb_id__tasks_delete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/tasks/import-from-bindings:
    post:
      tags:
      - ingestion_tasks
      summary: Create Import Task From Bindings
      operationId: create_import_task_from_bindings_kbs__kb_id__tasks_import_from_bindings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingTaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/tasks/reindex-from-bindings:
    post:
      tags:
      - ingestion_tasks
      summary: Create Reindex Task From Bindings
      operationId: create_reindex_task_from_bindings_kbs__kb_id__tasks_reindex_from_bindings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingTaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /kbs/{kb_id}/tasks/delete-from-bindings:
    post:
      tags:
      - ingestion_tasks
      summary: Create Delete Task From Bindings
      operationId: create_delete_task_from_bindings_kbs__kb_id__tasks_delete_from_bindings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: integer
          title: Kb Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingTaskCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks:
    get:
      tags:
      - ingestion_tasks
      summary: List Tasks
      operationId: list_tasks_tasks_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TaskResponse'
                type: array
                title: Response List Tasks Tasks Get
      security:
      - HTTPBearer: []
  /tasks/{task_id}:
    get:
      tags:
      - ingestion_tasks
      summary: Get Task
      operationId: get_task_tasks__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/process-pending:
    post:
      tags:
      - ingestion_tasks
      summary: Process Pending Tasks
      operationId: process_pending_tasks_tasks_process_pending_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Process Pending Tasks Tasks Process Pending Post
      security:
      - HTTPBearer: []
  /tasks/{task_id}/retry:
    post:
      tags:
      - ingestion_tasks
      summary: Retry Task
      operationId: retry_task_tasks__task_id__retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/cancel:
    post:
      tags:
      - ingestion_tasks
      summary: Cancel Task
      operationId: cancel_task_tasks__task_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/items:
    get:
      tags:
      - ingestion_tasks
      summary: Task Items
      operationId: task_items_tasks__task_id__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskItemRead'
                title: Response Task Items Tasks  Task Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BindingTaskCreateRequest:
      properties:
        binding_ids:
          items:
            type: integer
          type: array
          title: Binding Ids
      type: object
      title: BindingTaskCreateRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskCreateRequest:
      properties:
        source_paths:
          items:
            type: string
          type: array
          title: Source Paths
        credential_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Credential Id
      type: object
      required:
      - source_paths
      title: TaskCreateRequest
...
    TaskItemRead:
      properties:
        id:
          type: integer
          title: Id
        source_path:
          type: string
          title: Source Path
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        status:
          type: string
          title: Status
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        processed_chunks:
          anyOf:
          - type: integer
          - type: 'null'
          title: Processed Chunks
          default: 0
        source_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Version
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        error_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Type
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - source_path
      - status
      - created_at
      title: TaskItemRead
    TaskResponse:
      properties:
        id:
          type: integer
          title: Id
        owner_wallet_address:
          type: string
          title: Owner Wallet Address
        kb_id:
          type: integer
          title: Kb Id
        task_type:
          type: string
          title: Task Type
        status:
          type: string
          title: Status
        source_paths:
          items:
            type: string
          type: array
          title: Source Paths
        stats_json:
          type: object
          title: Stats Json
        error_message:
          type: string
          title: Error Message
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        queue_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Queue State
        queue_position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Queue Position
        current_running_task_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Current Running Task Id
        current_running_task_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Running Task Type
        cancelable:
          type: boolean
          title: Cancelable
          default: false
        claimed_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Claimed By
        heartbeat_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Heartbeat At
        last_stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Stage
        wait_duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wait Duration Ms
        run_duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Run Duration Ms
      type: object
      required:
      - id
      - owner_wallet_address
      - kb_id
      - task_type
      - status
      - source_paths
      - stats_json
      - error_message
      - created_at
      title: TaskResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://127.0.0.1:8000
  description: 本地开发环境