# knowledge/schemas/tasks.py
from __future__ import annotations

from datetime import datetime

from pydantic import BaseModel, Field


class TaskCreateRequest(BaseModel):
    source_paths: list[str]
    credential_id: int | None = None


class BindingTaskCreateRequest(BaseModel):
    binding_ids: list[int] = Field(default_factory=list)


class TaskResponse(BaseModel):
    id: int
    owner_wallet_address: str
    kb_id: int
    task_type: str
    status: str
    source_paths: list[str]
    stats_json: dict
    error_message: str
    created_at: datetime
    started_at: datetime | None = None
    finished_at: datetime | None = None
    queue_state: str | None = None
    queue_position: int | None = None
    current_running_task_id: int | None = None
    current_running_task_type: str | None = None
    cancelable: bool = False
    claimed_by: str | None = None
    heartbeat_at: datetime | None = None
    last_stage: str | None = None
    wait_duration_ms: int | None = None
    run_duration_ms: int | None = None

    model_config = {"from_attributes": True}


class TaskItemRead(BaseModel):
    id: int
    source_path: str
    file_name: str | None = None
    status: str
    message: str | None = None
    processed_chunks: int | None = 0
    source_version: str | None = None
    stage: str | None = None
    duration_ms: int | None = None
    error_type: str | None = None
    created_at: datetime